Split them further into 6 classes
